% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{onigomo_library_version}
\alias{onigomo_library_version}
\title{Return the version of the Onigmo library being used by the package}
\usage{
onigomo_library_version()
}
\description{
Return the version of the Onigmo library being used by the package
}
\examples{
onigomo_library_version()
}
