% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringore-package.R
\docType{package}
\name{stringore}
\alias{stringore}
\alias{stringore-package}
\title{Tidy Regular Expression Operations with Extensive Character Encoding Support}
\description{
The 'Onigmo' regular expression 'C' libary \url{https://github.com/k-takata/Onigmo} --
which is an extension of the 'Onigmura' \url{https://github.com/kkos/oniguruma} library --
provides support for a wide array of regular expression types and works across a
wide array of character encodings. Methods are provided to perform searching, matching
and extraction in a tidy (i.e. "data first") manner.
}
\details{
\href{https://github.com/jonclayden/ore}{ore} is another Onigmo-based R package for
handling regular expressions (it was the first to wrap this library). It abstracts
much of the complexity of Onigmo whereas the goal of \code{stringore} is to expose all
of the features and power of Onigmo while also providing a "tidy" abstraction layer.
}
\author{
Bob Rudis (bob@rud.is)
}
